webapp/requirements_streamlit.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.17.0
openpyxl>=3.0.10
python-dateutil>=2.8.2
gspread>=5.7.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
oauth2client>=4.1.3
python-dotenv>=1.0.0